backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp==2026.3.17
python-multipart==0.0.9
httpx>=0.27.0
openai>=1.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
aiosqlite>=0.19.0
stripe>=8.0.0
python-dotenv>=1.0.0
